[SC-65] refactor : scrap, like 패키지 정리
